breaking api changes (for Java client code only): - PackageInstaller and PackageUninstaller are interfaces now, so to use these classes one must obtain a singleton instance through getInstance() static method - callbacks are now inner interfaces of PackageInstaller and PackageUninstaller - in order to use PackageInstallerHelper methods a PackageInstaller instance must be passed as the first parameter
