+++ simpleinstaller/src/main/kotlin/io/github/solrudev/simpleinstaller/impl/InstallationEventsReceiver.kt
@@ -1,13 +1,14 @@
-package io.github.solrudev.simpleinstaller
+package io.github.solrudev.simpleinstaller.impl
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 
 internal class InstallationEventsReceiver : BroadcastReceiver() {
+
 	override fun onReceive(context: Context, intent: Intent) {
-		if (PackageInstaller.hasActiveSession) {
-			PackageInstaller.onPackageInstallerStatusChanged(context, intent)
+		if (PackageInstallerImpl.instance.hasActiveSession) {
+			PackageInstallerImpl.instance.onPackageInstallerStatusChanged(context, intent)
 		}
 	}
 }
